refactor: enforce style guidelines
